@@ -1,61 +1,32 @@
 {
     "name": "laravel/laravel",
     "type": "project",
-    "description": "Laravel Auth",
+    "description": "The Laravel Framework.",
     "keywords": [
         "framework",
         "laravel"
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
-        "ext-intl": "*",
-        "creativeorange/gravatar": "^1.0",
-        "doctrine/dbal": "^2.10",
+        "php": "^7.3",
         "fideloper/proxy": "^4.2",
-        "fruitcake/laravel-cors": "^1.0",
-        "google/recaptcha": "^1.2",
-        "guzzlehttp/guzzle": "^6.3",
-        "intervention/image": "^2.5",
-        "jaybizzle/laravel-crawler-detect": "^1.2",
-        "jeremykenedy/laravel-blocker": "^2.4",
-        "jeremykenedy/laravel-exception-notifier": "1.2",
-        "jeremykenedy/laravel-https": "^1.2",
-        "jeremykenedy/laravel-logger": "^3.6",
-        "jeremykenedy/laravel-phpinfo": "^1.2",
-        "jeremykenedy/laravel-roles": "^3.1",
-        "jeremykenedy/laravel2step": "^1.8",
-        "jeremykenedy/uuid": "^3.1",
-        "laravel/framework": "^7.0",
-        "laravel/helpers": "^1.2",
-        "laravel/socialite": "^4.3",
+        "fruitcake/laravel-cors": "^2.0",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "laravel/framework": "^8.0",
+        "laravel/jetstream": "^1.3",
+        "laravel/sanctum": "^2.6",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0",
-        "laravelcollective/html": "^6.1",
-        "pusher/pusher-php-server": "^4.1",
-        "rap2hpoutre/laravel-log-viewer": "^1.4",
-        "shvetsgroup/laravel-email-database-log": "^1.4",
-        "socialiteproviders/37signals": "^3.0",
-        "socialiteproviders/generators": "^4.6",
-        "socialiteproviders/instagram": "^3.0",
-        "socialiteproviders/linkedin": "^3.1",
-        "socialiteproviders/twitch": "^5.1",
-        "socialiteproviders/youtube": "^3.0"
+        "livewire/livewire": "^2.0",
+        "spatie/laravel-permission": "^3.17"
     },
     "require-dev": {
-        "barryvdh/laravel-debugbar": "^3.2",
-        "facade/ignition": "^2.0",
+        "barryvdh/laravel-debugbar": "^3.5",
+        "facade/ignition": "^2.3.6",
         "fzaninotto/faker": "^1.9.1",
         "mockery/mockery": "^1.3.1",
-        "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5"
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.3"
     },
-    "repositories": [
-        {
-            "type": "vcs",
-            "url":  "https://github.com/jeremykenedy/Generators-1.git"
-        }
-    ],
     "config": {
         "optimize-autoloader": true,
         "preferred-install": "dist",
@@ -67,12 +38,10 @@
         }
     },
     "autoload": {
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ],
         "psr-4": {
-            "App\\": "app/"
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
         }
     },
     "autoload-dev": {
