@@ -14,7 +14,7 @@ return [
     */
 
     'defaults' => [
-        'guard'     => 'web',
+        'guard' => 'web',
         'passwords' => 'users',
     ],
 
@@ -37,23 +37,14 @@ return [
 
     'guards' => [
         'web' => [
-            'driver'   => 'session',
+            'driver' => 'session',
             'provider' => 'users',
         ],
 
         'api' => [
-            'driver'   => 'token',
+            'driver' => 'token',
             'provider' => 'users',
-        ],
-
-        'user' => [
-            'driver'   => 'session',
-            'provider' => 'users',
-        ],
-
-        'administrator' => [
-            'driver'   => 'session',
-            'provider' => 'users',
+            'hash' => false,
         ],
     ],
 
@@ -77,7 +68,7 @@ return [
     'providers' => [
         'users' => [
             'driver' => 'eloquent',
-            'model'  => App\Models\User::class,
+            'model' => App\Models\User::class,
         ],
 
         // 'users' => [
@@ -103,10 +94,10 @@ return [
 
     'passwords' => [
         'users' => [
-            'provider'  => 'users',
-            'table'     => 'password_resets',
-            'expire'    => 60,
-            'throttle'  => 60,
+            'provider' => 'users',
+            'table' => 'password_resets',
+            'expire' => 60,
+            'throttle' => 60,
         ],
     ],
 
