@@ -8,85 +8,26 @@ return [
     |--------------------------------------------------------------------------
     |
     | This file is for storing the credentials for third party services such
-    | as Stripe, Mailgun, SparkPost and others. This file provides a sane
-    | default location for this type of information, allowing packages
-    | to have a conventional place to find your various credentials.
+    | as Mailgun, Postmark, AWS and more. This file provides the de facto
+    | location for this type of information, allowing packages to have
+    | a conventional file to locate the various service credentials.
     |
     */
 
     'mailgun' => [
         'domain' => env('MAILGUN_DOMAIN'),
         'secret' => env('MAILGUN_SECRET'),
+        'endpoint' => env('MAILGUN_ENDPOINT', 'api.mailgun.net'),
+    ],
+
+    'postmark' => [
+        'token' => env('POSTMARK_TOKEN'),
     ],
 
     'ses' => [
-        'key'    => env('SES_KEY'),
-        'secret' => env('SES_SECRET'),
-        'region' => 'us-east-1',
-    ],
-
-    'sparkpost' => [
-        'secret' => env('SPARKPOST_SECRET'),
-    ],
-
-    'stripe' => [
-        'model'  => App\Models\User::class,
-        'key'    => env('STRIPE_KEY'),
-        'secret' => env('STRIPE_SECRET'),
-    ],
-
-    'facebook' => [
-        'client_id'     => env('FB_ID'),
-        'client_secret' => env('FB_SECRET'),
-        'redirect'      => env('FB_REDIRECT'),
-    ],
-
-    'twitter' => [
-        'client_id'     => env('TW_ID'),
-        'client_secret' => env('TW_SECRET'),
-        'redirect'      => env('TW_REDIRECT'),
-    ],
-
-    'google' => [
-        'client_id'     => env('GOOGLE_ID'),
-        'client_secret' => env('GOOGLE_SECRET'),
-        'redirect'      => env('GOOGLE_REDIRECT'),
-    ],
-
-    'github' => [
-        'client_id'     => env('GITHUB_ID'),
-        'client_secret' => env('GITHUB_SECRET'),
-        'redirect'      => env('GITHUB_REDIRECT'),
-    ],
-
-    'youtube' => [
-        'client_id'     => env('YOUTUBE_KEY'),
-        'client_secret' => env('YOUTUBE_SECRET'),
-        'redirect'      => env('YOUTUBE_REDIRECT_URI'),
-    ],
-
-    'twitch' => [
-        'client_id'     => env('TWITCH_KEY'),
-        'client_secret' => env('TWITCH_SECRET'),
-        'redirect'      => env('TWITCH_REDIRECT_URI'),
-    ],
-
-    'instagram' => [
-        'client_id'     => env('INSTAGRAM_KEY'),
-        'client_secret' => env('INSTAGRAM_SECRET'),
-        'redirect'      => env('INSTAGRAM_REDIRECT_URI'),
-    ],
-
-    '37signals' => [
-        'client_id'     => env('37SIGNALS_KEY'),
-        'client_secret' => env('37SIGNALS_SECRET'),
-        'redirect'      => env('37SIGNALS_REDIRECT_URI'),
-    ],
-
-    'linkedin' => [
-        'client_id'     => env('LINKEDIN_KEY'),
-        'client_secret' => env('LINKEDIN_SECRET'),
-        'redirect'      => env('LINKEDIN_REDIRECT_URI'),
+        'key' => env('AWS_ACCESS_KEY_ID'),
+        'secret' => env('AWS_SECRET_ACCESS_KEY'),
+        'region' => env('AWS_DEFAULT_REGION', 'us-east-1'),
     ],
 
 ];
