@@ -14,24 +14,15 @@ class CreateUsersTable extends Migration
     public function up()
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('name')->unique();
-            $table->string('first_name')->nullable();
-            $table->string('last_name')->nullable();
-            $table->string('email')->unique()->nullable();
+            $table->id();
+            $table->string('name');
+            $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->rememberToken();
-            $table->boolean('activated')->default(false);
-            $table->string('token');
-            $table->ipAddress('signup_ip_address')->nullable();
-            $table->ipAddress('signup_confirmation_ip_address')->nullable();
-            $table->ipAddress('signup_sm_ip_address')->nullable();
-            $table->ipAddress('admin_ip_address')->nullable();
-            $table->ipAddress('updated_ip_address')->nullable();
-            $table->ipAddress('deleted_ip_address')->nullable();
+            $table->foreignId('current_team_id')->nullable();
+            $table->text('profile_photo_path')->nullable();
             $table->timestamps();
-            $table->softDeletes();
         });
     }
 
